hw/cxl: Update CXL Fixed Memory Window ACPI description to include Back Invalidate support.
Defaults for these windows has always been to enable anything QEMU supports.
With the addition of back invalidate support it is necessary to specify that
host windows support this.

CXL emulation is currently only suitable for software stack verification.
The relaxation of the restrictions on this window to include BI have no
affect on the OS until BI capable devices are added and until now these
have not existed. As such no backwards compatibility impacts are expected
from this change.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>